requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
sqlalchemy>=2.0.28
flask-migrate==4.0.5
flask-login==0.6.3
# Commented out PostgreSQL for local development (using SQLite instead)
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-wtf==1.2.1
email-validator==2.1.0
werkzeug==2.3.7
flask-bcrypt==1.0.1
python-dateutil==2.8.2
# Commented out packages requiring compilation which might cause issues on Windows
# numpy==1.26.2
# pandas==2.1.4
alembic==1.13.1
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.3
